dashboard/requirements.txt
streamlit
pandas
numpy
psycopg2-binary
altair
plotly
